+++ time64.c
@@ -874,7 +874,7 @@ char *ctime64( const Time64_T* time ) {
 void
 yd_from_ymd(struct TM *tm) {
     int i = 0;
-    const int year = tm->tm_year + 1900;
+    const int year = (int) tm->tm_year + 1900;
 		const int *dim = days_in_month[IS_LEAP(year)];
     tm->tm_yday = tm->tm_mday - 1;
     for (i = 0; i < tm->tm_mon; i++) {
@@ -901,7 +901,7 @@ leaps_thru_end_of(const int y) {
 }
 void
 ymd_from_yd(struct TM *tm) {
-		const int year = tm->tm_year + 1900;
+		const int year = (int) tm->tm_year + 1900;
 		const int *dim = days_in_month[IS_LEAP(year)];
     int days = tm->tm_yday;
     // Week Day
